pyproject.toml
[project]
authors = [{ name = "Harshal Laheri", email = "harshal@harshallaheri.me" }]
dependencies = [
    "advanced-alchemy[obstore]>=1.6.3",
    "beautifulsoup4>=4.14.0",
    "crawl4ai>=0.7.4",
    "google-genai>=1.39.1",
    "httpx>=0.28.1",
    "litestar-vite>=0.14.0",
    "litestar[standard]>=2.17.0",
    "psycopg[binary,pool]>=3.2.10",
    "python-dotenv>=1.1.1",
    "sqlspec>=0.26.0",
]
description = "Add your description here"
name = "litegen"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"


[tool.ruff]
line-length = 120
lint.ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "CPY001",
    "RUF012",
    "RUF029",
    "ERA001",
    "PLR0913",
    "PLR0917",
    "PLR6301",
    "PLR1702",
    "ANN204",
    "PLC0415",
    "FBT001",
    "FBT002",
    "TRY003",
    "EM101",
    "BLE001",
]

[dependency-groups]
dev = ["pre-commit>=4.3.0"]
